git-commit during version bumping automatically
This is currently broken because npm will not automatically commit/tag
if the package is not at the repo root.

This project is using lerna, but we don't need lerna since we're only
modifying a single subpackage. So instead, we hook into the version
lifecycle event to do the git committing ourselves. (But no tagging;
there are still multiple packages in this repo after all.)