Fix sprintf for undefined format strings
Since we remove all the ERROR_MESSAGES and LOG_MESSAGES format strings,
the sprintf format function blows up when the format string is
undefined. This proxy allows us to return a dummy format string
regardless which "getter" property is accessed on those message objects.

The dummy format string returned uses three placeholders; which should
cover nearly all (if not all) possible invocations of the sprintf
function (having a placeholder for up to 3 arguments). This way the
sprintf function will basically operate as a "to-string" on all the
arguments provided.